requirements.txt
ansio==0.0.1
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-openai==0.1.15
langchain-community==0.2.7
langchain-anthropic==0.1.19
langchain-google-genai==1.0.7
webcolors==24.6.0
sentence-transformers==3.0.1
docker==7.1.0
paramiko==3.4.0
duckduckgo_search==6.1.12
inputimeout==1.0.4
pynput==1.7.7
pypdf==4.3.1
Flask[async]==3.0.3
Flask-BasicAuth==0.2.0
faiss-cpu==1.8.0.post1
playwright==1.46.0